Update solution - CodeHub

+++ 7-kyu/most-digits/most-digits.java
@@ -7,7 +7,7 @@ public class MostDigits {
         result = numbers[i];
       }
     }
-      return (int)result;
+      return result;
   }
 ​
 }
